requirements.txt
PyQt6==6.6.1
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
pandas==2.1.4